fix(66-2): split CEF helpers into separate files per one-export lint rule
- Rename fidelity-cef-universe.helper.ts -> build-cef-classification.helper.ts
- Create create-universe-entry.helper.ts as separate file
- Fix no-anonymous-functions lint errors: replace arrow fns with named fns
- Update fidelity-data-mapper.function.ts imports
