+++ apps/server/src/app/routes/import/fidelity-data-mapper.function.ts
@@ -1,8 +1,6 @@
 import { prisma } from '../../prisma/prisma-client';
-import {
-  buildCefClassification,
-  createUniverseEntry,
-} from './fidelity-cef-universe.helper';
+import { buildCefClassification } from './build-cef-classification.helper';
+import { createUniverseEntry } from './create-universe-entry.helper';
 import { FidelityCsvRow } from './fidelity-csv-row.interface';
 import { isInLieuRow } from './is-in-lieu-row.function';
 import { isSplitFromRow } from './is-split-from-row.function';
